+++ inst/app/modules/news_ui.R
@@ -17,8 +17,10 @@ newsUI <- function(id) {
         <h3>2025-12-XX</h3>
         
         The v1.2.0 release of RMAVIS incorporates the Minnesota Native Plant Community Classification (MNNPC).
-        This has involved refactoring several modules to ensure that RMAVIS is extensible and can easily incorporate additional vegetation classifications,
-        however this release contains no breaking changes for the GB-NVC.
+        This has involved refactoring several modules to ensure that RMAVIS is extensible and can easily incorporate additional vegetation classifications.
+        A second major change is the methodology used to calculate Diversity measures, specifically RMAVIS now utilises the approach described in Reeve et al (2014) *How to partition diversity* 
+        as implemented in the R package {rdiversity}; for more information please see the documentation.
+        This release contains no breaking changes for matching vegetation plots to the GB-NVC.
        
 
         <h3>2025-09-11</h3>